+++ src/tools/index.ts
@@ -2,6 +2,7 @@
  *  Tool Module Exports
  *--------------------------------------------------------------------------------------------*/
 
+export { DelegateToAgentTool } from "./delegateToAgent";
 export { MiniMaxSearchTool } from "./minimaxSearch";
 export { registerAllTools } from "./registry";
 export { ZhipuSearchTool } from "./zhipuSearch";